requirements.txt
djangorestframework-simplejwt==4.3.0
PyJWT==1.7.1
dj-rest-auth==2.1.3
django-allauth==0.44.0
